Fix
crud user , luồng bán hàng  thêm trạng 2 trạng thái đơn hàng , sửa trạng thái xác nhận đơn hàng không trừ trong db

+++ DATN_BE/src/main/java/com/example/datn_be/dto/OrderDetailsDTO.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class OrderDetailsDTO {
 
     private Integer orderDetailId;
-    private BigDecimal totals;
+    private Integer totals;
     private String orderCode;
     private BigDecimal amount;
     private String name;
